+++ bogenliga/src/app/modules/shared/components/buttons/shortcut-button/shortcut-button.component.html
@@ -4,7 +4,12 @@
     <ng-container *ngFor="let button of config.shortcutButtons">
       <div *ngIf="hasUserPermissions(button.permissions) && hasUserRole(button.roles)" class="buttonwrapper">
         <div class="Button">
-          <button type="button" class="shortcut-btn" [routerLink]="button.route" routerLinkActive="active">
+          <button *ngIf="button.title !== 'Mannschaftsmitglied hinzufügen'"
+                  type="button" class="shortcut-btn" [routerLink]="button.route" routerLinkActive="active">
+            <img class="buttonlogo" *ngIf="button.icon" [src]="button.icon" alt="{{ button.title }}" />
+          </button>
+          <button *ngIf="button.title === 'Mannschaftsmitglied hinzufügen'"
+                  type="button" class="shortcut-btn" [routerLink]="getCorrectLink()" routerLinkActive="active">
             <img class="buttonlogo" *ngIf="button.icon" [src]="button.icon" alt="{{ button.title }}" />
           </button>
           <ng-container>
